requirements.txt
streamlit
mysql-connector-python
matplotlib
pillow
streamlit-option-menu
